api/llm/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
openai>=1.0.0
python-dotenv==1.0.0
pydantic==2.4.2
starlette==0.27.0 